Make zeroconfig more forgiving of data in autodl.cfg
In the wild I've now seen a handful of configuration files which
do not conform to PHP's strict ini parsing. As such I concluded
it may be best to use a more lax parsing mode.

I've also changed it to only accept the password and port from
the options section, as I've seen configuration files in the wild
with different passwords and port values defined under random
sections within the ini file.